Allow running other Gradle tasks
