+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/RunBuild.kt
@@ -26,112 +26,12 @@
 
 package io.spine.internal.gradle
 
-import java.io.File
-import java.util.concurrent.TimeUnit
-import org.gradle.api.DefaultTask
-import org.gradle.api.GradleException
-import org.gradle.api.tasks.Internal
-import org.gradle.api.tasks.TaskAction
-import org.gradle.internal.os.OperatingSystem
-
 /**
- * A Gradle task which runs another Gradle build.
- *
- * Launches Gradle wrapper under a given [directory] with the `build` task. The `clean` task is also
- * run if current build includes a `clean` task.
- *
- * The build writes verbose log into `$directory/build/debug-out.txt`. The error output is written
- * into `$directory/build/error-out.txt`.
+ * Runs the `build` task via Gradle Wrapper.
  */
-@Suppress("unused")
-open class RunBuild : DefaultTask() {
+open class RunBuild : RunGradle() {
 
-    /**
-     * Path to the directory which contains a Gradle wrapper script.
-     */
-    @Internal
-    lateinit var directory: String
-
-    /**
-     * Names of Gradle properties to copy into the launched build.
-     *
-     * The properties are looked up in the root project. If a property is not found, it is ignored.
-     *
-     * See [Gradle doc](https://docs.gradle.org/current/userguide/build_environment.html#sec:gradle_configuration_properties)
-     * for more info about Gradle properties.
-     */
-    @Internal
-    var includeGradleProperties: MutableSet<String> = mutableSetOf()
-
-    @TaskAction
-    private fun execute() {
-        // Ensure build error output log.
-        // Since we're executing this task in another process, we redirect error output to
-        // the file under the `build` directory.
-        val buildDir = File(directory, "build")
-        if (!buildDir.exists()) {
-            buildDir.mkdir()
-        }
-        val errorOut = File(buildDir, "error-out.txt")
-        val debugOut = File(buildDir, "debug-out.txt")
-
-        val command = buildCommand()
-        val process = startProcess(command, errorOut, debugOut)
-
-        /*  The timeout is set because of Gradle process execution under Windows.
-            See the following locations for details:
-              https://github.com/gradle/gradle/pull/8467#issuecomment-498374289
-              https://github.com/gradle/gradle/issues/3987
-              https://discuss.gradle.org/t/weirdness-in-gradle-exec-on-windows/13660/6
-         */
-        val completed = process.waitFor(10, TimeUnit.MINUTES)
-        val exitCode = process.exitValue()
-        if (!completed || exitCode != 0) {
-            val errorOutExists = errorOut.exists()
-            if (errorOutExists) {
-                logger.error(errorOut.readText())
-            }
-            throw GradleException("Child build process FAILED." +
-                    " Exit code: $exitCode." +
-                    if (errorOutExists) " See $errorOut for details."
-                    else " $errorOut file was not created."
-            )
-        }
+    init {
+        task("build")
     }
-
-    private fun buildCommand(): List<String> {
-        val script = buildScript()
-        val command = mutableListOf<String>()
-        command.add("${project.rootDir}/$script")
-        val shouldClean = project.gradle
-            .taskGraph
-            .hasTask(":clean")
-        if (shouldClean) {
-            command.add("clean")
-        }
-        command.add("build")
-        command.add("--console=plain")
-        command.add("--debug")
-        command.add("--stacktrace")
-        command.add("--no-daemon")
-        val rootProject = project.rootProject
-        includeGradleProperties
-            .filter { rootProject.hasProperty(it) }
-            .map { name -> name to rootProject.property(name).toString() }
-            .forEach { (name, value) -> command.add("-P$name=$value") }
-        return command
-    }
-
-    private fun buildScript(): String {
-        val runsOnWindows = OperatingSystem.current().isWindows()
-        return if (runsOnWindows) "gradlew.bat" else "gradlew"
-    }
-
-    private fun startProcess(command: List<String>, errorOut: File, debugOut: File) =
-        ProcessBuilder()
-            .command(command)
-            .directory(project.file(directory))
-            .redirectError(errorOut)
-            .redirectOutput(debugOut)
-            .start()
 }
